Added RegisterEntries file
