+++ Client/Clique/Common/Models/RegisterEntries.swift
@@ -1,0 +1,16 @@
+//
+//  RegisterEntries.swift
+//  Clique
+//
+//  Created by Infinum on 26.11.2022..
+//
+
+import Foundation
+
+struct RegisterEntries : Codable {
+
+    let email : String
+    let password : String
+    let name: String
+    let surname: String
+}
